+++ src/hooks/useCanGrant.ts
@@ -1,20 +1,17 @@
 import { StreamPermission } from 'streamr-client'
-import useCurrentAbility from '$/hooks/useCurrentAbility'
-import useLoadCurrentAbilityEffect from '$/hooks/useLoadCurrentAbilityEffect'
+import useAbility from '$/hooks/useAbility'
 import { useSelectedRoomId } from '$/features/room/hooks'
 import { useWalletAccount } from '$/features/wallet/hooks'
 import useJustInvited from './useJustInvited'
 
 export default function useCanGrant() {
-    const canGrant = useCurrentAbility(StreamPermission.GRANT)
-
     const address = useWalletAccount()
 
     const selectedRoomId = useSelectedRoomId()
 
-    const justInvited = useJustInvited(selectedRoomId, address)
+    const canGrant = useAbility(selectedRoomId, address, StreamPermission.GRANT)
 
-    useLoadCurrentAbilityEffect(StreamPermission.GRANT)
+    const justInvited = useJustInvited(selectedRoomId, address)
 
     return canGrant && !justInvited
 }
